cli/pyproject.toml
[tool.poetry]
name = "peach"
version = "0.1.0"
description = ""
authors = ["Sami Junior <sami.juniorco@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
sounddevice = "^0.4.7"
soundfile = "^0.12.1"
openai = "^1.33.0"
pvporcupine = "^3.0.2"
groq = "^0.8.0"
serpapi = "^0.1.5"
google-search-results = "^2.4.2"
python-dotenv = "^1.0.1"
pvrecorder = "^1.2.2"
elevenlabs = "^1.3.0"
flask = "^3.0.3"
flask-cors = "^4.0.1"
tavily-python = "^0.3.3"
deepgram-sdk = "^3.3.1"
pydub = "^0.25.1"
simpleaudio = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
